Split import service into two distinct files
