+++ src/app/core/utils/beneficiaries-import.service.ts
@@ -1,0 +1,48 @@
+import { Injectable } from '@angular/core';
+import { map } from 'rxjs/operators';
+import { Project } from 'src/app/model/project';
+import { HouseholdsService } from '../api/households.service';
+
+
+@Injectable({
+    providedIn: 'root'
+})
+export class ImportService {
+
+    private email: string;
+    // Token stored on file upload and reused for future steps
+    private token: string;
+    // Current project ( Todo: store it as Project object )
+    private project: Project;
+
+    private csvFile: File;
+
+    constructor(
+        private householdsService: HouseholdsService,
+    ) {}
+
+    setImportContext(email: string, project: Project, csvFile: File ) {
+        this.email = email;
+        this.project = project;
+        this.csvFile = csvFile;
+    }
+
+    // Sends CSV to server and sets variables needed for the rest of the import session
+    sendCsv() {
+        this.token = undefined;
+        const body = new FormData();
+        body.append('file', this.csvFile);
+
+        return this.sendStepUserData(body);
+    }
+
+    sendStepUserData(data: any) {
+        return this.householdsService.sendDataToValidation(this.email, data, this.project.id, this.token).pipe(
+            map((response: Response) => {
+                this.token = response['token'];
+                delete response['token'];
+                return response;
+            })
+        );
+    }
+}
